feat(tests): extract reusable chat integration base
